math: fix sinhf overflow boundary; verify sinhf/coshf/tanhf correctly rounded
expm1f_ compared its overflow threshold as a float, which rounds up to
0x1.62e43p+9.  The boundary input x = 0x1.62e43p+9 then slipped past the
strict `>` into the shift_ path, where n = 1024 overflows the double exponent
field to NaN -- so sinhf(+-0x1.62e43p+9) returned +-NaN instead of +-inf.
Compare in double (as expf_ already does) so that input takes the overflow
path.

coshf and tanhf were already correctly rounded; add exhaustive 2^32 oracle
sweeps for all three.

+++ src/math/float/expm1f.h
@@ -7,8 +7,12 @@
 
 static double expm1f_(float x)
 {
-    const float minimum = -37.42994775023704671;
-    const float maximum = 709.78271289338399684;
+    /* Compare in double: a float threshold for `maximum` rounds up to
+     * 0x1.62e43p+9, letting that input slip past into the shift_ path where the
+     * exponent (n = 1024) overflows to NaN.  expf_ compares in double for the
+     * same reason. */
+    const double minimum = -37.42994775023704671;
+    const double maximum = 709.78271289338399684;
 
     const float log2e = 1.442695040888963407;
     const double ln2 = 0.6931471805599453094;
